sample_project/requirements.txt
# Sample project with known vulnerable packages (Python 3.14 compatible)
# These have known CVEs for testing purposes

requests==2.32.4
flask==3.0.0
jinja2==3.1.6
werkzeug==3.1.5
certifi==2024.7.4
pyyaml==6.0.1
